Feature/formatint (#57)
* add formatint and handle negative int

* started testing

* int tests

* last tests added

* restored updated date as a test

* pretty please merge

* format int changed for new stuff

* small fixes

* changed flag order in Makefile

for better compatibility with other systems.
(on Debian in WSL on Windows it won't compile if the flags are at the end)

* small fix for include paths

* added i to dispatcher

* restructured formatint funcs for lenmod

ft_formatint is now updated for the new workflow with ft_formatdecimal,
This is so we can integrate lenmod. ft_handle_negint is now ft_format_negint.
That function is also updated, it is now an alternative to ft_formatint instead of a part of.

* added ft_format_decimal and PLACEHOLDERS

added ft_formatdecimal which will dispatch `i` and `d` conversions,
depending on the length modifier provided.
also added placeholders that need to be build for lenmod conversions.

* small change to support larger numbers

* added ft_write_flags

this function will write flags when needed.

* modified header for new and modified functions.

* added and updated tests for int

added some tests, mostly integration tests calling the ft_printf function.

* added new functions to sources file

* small fixes from merge

* added negative test with - flag

* made ft_write_flags modify len properly.

+++ libft/ft_nbrlenbase.c
@@ -6,13 +6,13 @@
 /*   By: lgutter <lgutter@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/02/02 12:11:46 by lgutter        #+#    #+#                */
-/*   Updated: 2019/09/30 18:05:24 by ivan-tey      ########   odam.nl         */
+/*   Updated: 2019/11/13 13:10:30 by lgutter       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t		ft_nbrlenbase(long number, int base)
+size_t		ft_nbrlenbase(long long number, int base)
 {
 	size_t len;
